[Fix] : 유저 뷰모델 수정 및 적용 #5
